Fix #29
Edit code so only player cna roll dice

Only player has attack and defend bonus added to its attack and defend value
