+++ src/main/java/Game/Characters/Player.java
@@ -12,6 +12,8 @@ import Game.Currency.Gold;
  */
 public class Player extends Character {
     private Gold gold;
+    private int attackBonus;
+    private int defenseBonus;
 
     private static final String playerModel =
             "      __      _\n" +
@@ -61,10 +63,38 @@ public class Player extends Character {
     @Override
     public int getDefenseValue() {
         if (isDefending) {
-            return defenseValue * 3;
+            return (defenseValue + defenseBonus) * 3;
         }
-        return defenseValue;
+        return defenseValue + defenseBonus;
     }
+
+    /**
+     * Return the character's attack value.
+     *
+     * @return an integer representing the character's attack value.
+     */
+    public int getAttackValue() {
+        return attackValue + attackBonus;
+    }
+
+    /**
+     * Set player's attack bonus value.
+     *
+     * @param bonus dice outcome for attack action.
+     */
+    public void setAttackBonus(int bonus){
+        attackBonus = bonus;
+    }
+
+    /**
+     * Set player's defend bonus value.
+     *
+     * @param bonus dice outcome for defend action.
+     */
+    public void setDefenseBonus(int bonus){
+        defenseBonus = bonus;
+    }
+
 
     @Override
     public String toString() {
