+++ src/main/java/Game/Events/Battle/PlayerTurn.java
@@ -7,6 +7,7 @@ import Game.Actions.BattleAction.AttackAction;
 import Game.Actions.BattleAction.DefendAction;
 import Game.Actions.BattleAction.FleeAction;
 import Game.Characters.Character;
+import Game.Characters.Player;
 import Game.RollDice;
 
 /**
@@ -47,24 +48,25 @@ public class PlayerTurn extends Turn {
      *
      */
     public void handleAction() {
+        ((Player) player).setAttackBonus(0);
+        ((Player) player).setDefenseBonus(0);
+
         Action currentAction = getCurrAction();
         int diceValue = RollDice.rollDice();
         int diceOutcome = RollDice.diceOutcome(diceValue);
 
         if (currentAction instanceof AttackAction) {
-            player.setAttackBonus(diceOutcome);
+            ((Player) player).setAttackBonus(diceOutcome);
             player.attack(enemy);
         }
         else if (currentAction instanceof DefendAction) {
-            player.setDefenseBonus(diceOutcome);
+            ((Player) player).setDefenseBonus(diceOutcome);
             player.setDefending(true);
+            //System.out.println("defense, defenseBonus" +((Player) player).getDefenseValue()+" "+((Player) player).defenseBonus );
         }
         else if (currentAction instanceof FleeAction) {
             hasSurrendered = true;
         }
-
-        player.setAttackBonus(0);
-        player.setDefenseBonus(0);
 
     }
